09._Environment_Variables/02._python/main.py
from dotenv import load_dotenv
import os

load_dotenv()
print(os.getenv("API_KEY"))
print(os.getenv("PORT"))

print("\nOR:\n")

from dotenv import dotenv_values

config = dotenv_values(".env")
print(config["API_KEY"])
print(config["PORT"])
