09._Environment_Variables/02._python/pyproject.toml
[project]
name = "02-python"
version = "0.1"
description = "Working with environment variables in python"
authors = [
    {name = "VictorHanert",email = "vict660m@stud.kea.dk"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "02"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
